Add camera status endpoint.  Add camera process check for trying to stop app

+++ src/security/security.service.ts
@@ -1,20 +1,18 @@
-// const childProcess: ChildProcess = require('child_process');
 const childProcess = require('child_process');
 
 import { Injectable } from '@nestjs/common';
-// import { ChildProcess } from 'child_process';
 
 @Injectable()
 export class SecurityService {
 
-  private securityAppProcess;
-  private pathToHomeWatcherCameraApp
+  public securityAppProcess = null;
+  private pathToHomeWatcherCameraApp = `/home/pi/Home-Watcher-remake-PI/app.js`;
 
   startSecurity(): boolean {
     try {
       console.log('Starting security app');
 
-      this.securityAppProcess = childProcess.fork(`/home/pi/Home-Watcher-remake-PI/app.js`);
+      this.securityAppProcess = childProcess.fork(this.pathToHomeWatcherCameraApp);
 
       this.securityAppProcess.on('exit', function (code, signal) {
         console.log(`Camera child process exited with code ${code} and signal ${signal}`);
@@ -35,19 +33,24 @@ export class SecurityService {
 
   stopSecurity(): boolean {
     try {
+      if(this.securityAppProcess) {
+        console.log('Stopping security app');
 
-      console.log('Stopping security app');
+        // https://nodejs.dev/how-to-exit-from-a-nodejs-program
+  
+        // process.kill(process.pid, 'SIGTERM')
+  
+        this.securityAppProcess.kill();
+        this.securityAppProcess = null;
+  
+        console.log('Security app has successfully been stopped');
 
-      // https://nodejs.dev/how-to-exit-from-a-nodejs-program
+        return true;
+      }
 
-      // process.kill(process.pid, 'SIGTERM')
-
-      this.securityAppProcess.kill();
-      this.securityAppProcess = null;
-
-      console.log('Security app has successfully been stopped');
-
-      return true;
+      console.log('Security app is not running');
+      return false;
+      
     } catch (error) {
       console.log('error', error);
       return false;
